Создал компонент и вынес кнопку logout в профиль
