+++ packages/client/src/entities/User/ui/ProfileForm/ui/Logout/index.ts
@@ -1,0 +1,3 @@
+import { Logout } from './ui/Logout'
+
+export { Logout as LogoutBtn }
